create httpPing and using it instead of icmp ping

+++ migration/createTable.sql
@@ -7,7 +7,7 @@ CREATE TABLE results(
 	id SERIAL PRIMARY KEY,
 	destination VARCHAR(255),
 	status VARCHAR(30),
-	time smallint,
+	time integer,
 	created_at TIMESTAMP WITH TIME ZONE DEFAULT NOW()
 );
